Only need repo info to get lanes
